requirements.txt
pymongo 
langchain 
langchainhub 
langchain-groq
python-dotenv
streamlit